refactor: renderer2 + ssr refactoring
