+++ .gitignore
@@ -39,8 +39,12 @@ testem.log
 .DS_Store
 Thumbs.db
 
-*.map
-*.metadata.json
-*.d.ts
+index.metadata.json
+index.d.ts
 index.js
-angular2-drag-scroll.js
+index.js.map
+
+/src/**/*.d.ts
+/src/**/*.js
+/src/**/*.js.map
+/src/**/*.metadata.json
