+++ demo/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AppComponent } from './app.component';
+
 import { HomeComponent } from './home/home.component';
 import { NotFoundComponent } from './not-found/not-found.component';
 
@@ -11,7 +11,7 @@ export const routes: Routes = [
     pathMatch: 'full'
   },
   {
-    path:'404',
+    path: '404',
     component: NotFoundComponent
   },
   {
