+++ demo/app/app.module.ts
@@ -1,14 +1,15 @@
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { MaterialModule } from '@angular/material';
+import { MdButtonModule, MdSlideToggleModule, MdToolbarModule } from '@angular/material';
 
-import 'style-loader!@angular/material/prebuilt-themes/indigo-pink.css';
+// import 'style-loader!@angular/material/prebuilt-themes/indigo-pink.css';
 
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
-import { DragScrollModule } from '../../src/angular2-drag-scroll';
+import { DragScrollModule } from '../../src/index';
 import { NotFoundComponent } from './not-found/not-found.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HomeComponent } from './home/home.component';
@@ -27,12 +28,15 @@ import { GithubComponent } from './github/github.component';
     HomeComponent
   ],
   imports: [
-    AppRoutingModule,
-    DragScrollModule,
-    BrowserModule,
+    BrowserModule.withServerTransition({appId: 'drag-scroll-app'}),
+    BrowserAnimationsModule,
     FormsModule,
     HttpModule,
-    MaterialModule
+    AppRoutingModule,
+    DragScrollModule,
+    MdToolbarModule,
+    MdButtonModule,
+    MdSlideToggleModule
   ],
   providers: [
   ],
