+++ demo/main.server.ts
@@ -1,0 +1,1 @@
+export { AppServerModule } from './app/app-server.module';
