+++ demo/main.ts
@@ -1,5 +1,6 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { enableProdMode } from '@angular/core';
+
 import { environment } from './environments/environment';
 import { AppModule } from './app/app.module';
