+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "angular2-drag-scroll",
-  "version": "1.4.0-rc.6",
+  "version": "2.0.0-alpha.0",
   "description": "Lightweight drag to scroll directive for Angular2",
   "contributors": [
     {
@@ -14,11 +14,14 @@
     "ng": "ng",
     "start": "ng serve",
     "test": "ng test --single-run",
-    "clean:build": "npm run rimraf -- **/*.js **/*.d.ts **/*.map index.js index.d.ts index.js.map",
+    "clean:build": "npm run rimraf -- src/**/*.js src/**/*.d.ts src/**/*.map src/**/*.metadata.json index.js index.d.ts index.js.map index.metadata.json",
     "clean": "npm run rimraf -- node_modules && npm cache clean",
     "rimraf": "rimraf",
     "ngc": "ngc -p src/tsconfig.json",
-    "tsc": "tsc -p src"
+    "tsc": "tsc -p src",
+    "build:all": "ng build --prod && ng build --prod --app 1",
+    "build:aot": "ng build --prod --aot && ng build --prod --aot --app 1",
+    "build:optimizer": "ng build --prod --aot --build-optimizer && ng build --prod --aot --build-optimizer --app 1"
   },
   "repository": {
     "type": "git",
@@ -36,41 +39,46 @@
   },
   "homepage": "https://github.com/bfwg/angular2-drag-scroll#readme",
   "peerDependencies": {
-    "@angular/common": "4.2.3",
-    "@angular/compiler": "4.2.3",
-    "@angular/core": "4.2.3"
+    "@angular/common": "4.3.6",
+    "@angular/compiler": "4.3.6",
+    "@angular/core": "4.3.6"
   },
   "devDependencies": {
-    "@angular/animations": "4.2.3",
-    "@angular/cli": "1.1.1",
-    "@angular/common": "4.2.3",
-    "@angular/compiler": "4.2.3",
-    "@angular/compiler-cli": "4.2.3",
-    "@angular/core": "4.2.3",
-    "@angular/forms": "4.2.3",
-    "@angular/http": "4.2.3",
-    "@angular/material": "2.0.0-beta.6",
-    "@angular/platform-browser": "4.2.3",
-    "@angular/platform-browser-dynamic": "4.2.3",
-    "@angular/router": "4.2.3",
+    "@angular/animations": "4.3.6",
+    "@angular/cdk": "2.0.0-beta.10",
+    "@angular/cli": "1.4.1",
+    "@angular/common": "4.3.6",
+    "@angular/compiler": "4.3.6",
+    "@angular/compiler-cli": "4.3.6",
+    "@angular/core": "4.3.6",
+    "@angular/forms": "4.3.6",
+    "@angular/http": "4.3.6",
+    "@angular/material": "2.0.0-beta.10",
+    "@angular/platform-browser": "4.3.6",
+    "@angular/platform-browser-dynamic": "4.3.6",
+    "@angular/platform-server": "4.3.6",
+    "@angular/router": "4.3.6",
+    "@nguniversal/express-engine": "^1.0.0-beta.3",
     "@types/core-js": "^0.9.43",
-    "@types/jasmine": "2.5.45",
-    "@types/node": "6.0.78",
-    "codelyzer": "3.0.1",
-    "core-js": "2.4.1",
+    "@types/jasmine": "2.5.53",
+    "@types/node": "6.0.88",
+    "codelyzer": "3.1.1",
+    "compression": "^1.7.0",
+    "core-js": "2.5.1",
+    "express": "^4.15.4",
     "jasmine-core": "2.6.4",
     "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.0",
+    "karma": "1.7.1",
     "karma-chrome-launcher": "2.1.1",
     "karma-cli": "1.0.1",
     "karma-coverage-istanbul-reporter": "1.3.0",
     "karma-jasmine": "1.1.0",
     "karma-jasmine-html-reporter": "0.2.2",
     "rimraf": "2.5.4",
-    "rxjs": "5.4.1",
-    "ts-node": "3.0.6",
+    "rxjs": "5.4.3",
+    "ts-node": "3.2.0",
     "tslint": "5.3.2",
     "typescript": "2.3.4",
-    "zone.js": "0.8.12"
+    "zone.js": "0.8.17"
   }
 }
