+++ src/drag-scroll.html
@@ -1,0 +1,3 @@
+<div class="drag-scroll-content" #contentRef>
+  <ng-content></ng-content>
+</div>
